keyword change - extended to csv2mysql
The keyword 'extended' was used at project inception to represent a custom LogFormat and is propagated throughout the application. Concerns were raised the use of this keyword could be confused with Extended Log Format. Decision was made to apply this change now while downloads are low since table and column names are involved.